Ensure terraform has been initialized before destroying
Prior to this commit, if a user tried to run destroy before doing a
provision, the plan would fail because terraform had not been
initalized. This commit adds the terraform initialize to the destroy
plan to allow the destroy plan to be functional even if there are no
deployed instances.