(feat) Allow for specifying the IP path from terraform
Prior to this commit, the URI in the terraform resolve_references was
hard coded. As different terraform configurations may have different
paths to the IP address, the code needs to be dynamic to handle the
differences. This commit allows for the terraform output to present an
`ip_path` output to specify the path to the IP address for the resource.
If it is not specified, the default will be used.