Provide option for selecting existing networks
Commit add the subnetwork and subnetwork_project parameters to enable
the attachment of PE to existing networks. To allow for these instance
to be attached the the network where bolt is being ran, a ssh_ip_mode
parameter was also added which allows for the selection of public or
private IP addresses.