Empty inventory.yaml to prevent failures
Bolt's inventory.yaml is limited in how dynamic is can be when combined
with the Terraform plugin, this limitation results in failures in some
deployment scenarios which are supported by pecdm::provision, e.g. those
which depend entirely on private IP addresses. This commit removes
references to the plugin, which were not required for provisioning to
function because the plan constructs its own inventory upon execution.

Example inventory files for each supported cloud can now be found in the
examples directory. These inventory files can be used in the event you
want to run non-pecdm Bolt plans and tasks against a successfully
provisioned cluster.