Internal restructuring to make more moduler
The provision and destroy plans broken up so that their individuals
steps can be ran independently of each. The ability to use each
component independently enables the module to be composed by the user is
new ways so they can address the additional steps required for their
infrastructure, e.g. custom DNS names or classification changes.

The default functionality of the original provision and destroy plans
remains.